Moved queries and mutations to queries.js file
