+++ src/pages/Groups.js
@@ -8,29 +8,12 @@ import { Text, Button, Card, Icon } from 'react-native-paper';
 import { styles } from '../styles/main-styles';
 
 import Loader from '../components/helpers/Loader';
+import { GET_GROUPS_QUERY } from '../models/Queries';
 
 export default function Groups({ navigation }) {
     // let [searchForNew, setSearchForNew] = useState(false);
 
-    const { data, errors, loading } = useQuery(gql`query GetGroups {
-        groups: getGroups(includeDetails: true) {
-            groupId
-            groupName
-            description
-            location
-            groupMembers {
-                username
-                firstName
-                lastName
-                role
-            }
-            owner {
-                username
-                firstName
-                lastName
-            }
-        }
-    }`);
+    const { data, errors, loading } = useQuery(GET_GROUPS_QUERY);
 
     if(loading) {
         return <Loader />
